Added api path '/order' to order the dissemination of information.
The DAWeb application needs a mix of basic authentication and cookie authentication (JSESSIONID) to accept the ordering of information. I didn't get this working using plain node.js functionality (due to an unknown reason the cookie was not set correctly when POSTing to '/automatedRetrieval/queueForRetrievalJSON') which is why I added the 'request' library.