Added http basic authentication.
The username and password was removed from the config file and http basic authentication was added.
But: Only the existence of a username and password is checked, the credentials are than forwarded to the actual danrw url. For the api pathes '/ingest' and '/disseminate' only the existence of the user folder is checked, no authentication is done for these two urls.